+++ modules/app-core/src/commonMain/kotlin/org/archivekeep/app/core/operations/addpush/AddAndPushOperationServiceImpl.kt
@@ -121,6 +121,7 @@ class AddAndPushOperationServiceImpl(
                                     dst = destinationRepo,
                                     base = repo,
                                     filename = fileToPush,
+                                    progressReport = { /* TODO */ },
                                 )
 
                                 println("copied: $destinationRepoID - $fileToPush")