+++ modules/files/src/main/kotlin/org/archivekeep/files/operations/sync/SyncSubOperation.kt
@@ -1,5 +1,7 @@
 package org.archivekeep.files.operations.sync
 
+import kotlinx.coroutines.flow.MutableStateFlow
+import org.archivekeep.files.operations.tasks.InProgressOperationStats
 import org.archivekeep.files.repo.Repo
 
 sealed interface SyncSubOperation {
@@ -7,5 +9,6 @@ sealed interface SyncSubOperation {
         base: Repo,
         dst: Repo,
         logger: SyncLogger,
+        inProgressOperationStatsMutableFlow: MutableStateFlow<List<InProgressOperationStats>>,
     )
 }