+++ modules/files/src/main/kotlin/org/archivekeep/files/operations/tasks/InProgressOperationStats.kt
@@ -1,0 +1,10 @@
+package org.archivekeep.files.operations.tasks
+
+import kotlin.time.Duration
+
+interface InProgressOperationStats {
+    val timeConsumed: Duration
+    val timeEstimated: Duration?
+    val velocity: Any
+    val completion: Float
+}