+++ modules/files/src/main/kotlin/org/archivekeep/testing/fixtures/FixtureRepo.kt
@@ -43,6 +43,7 @@ class FixtureRepo(
         filename: String,
         info: ArchiveFileInfo,
         stream: InputStream,
+        monitor: (copiedBytes: Long) -> Unit,
     ) {
         TODO("Not yet implemented")
     }