+++ modules/files/src/main/kotlin/org/archivekeep/testing/storage/SpeedLimitedRepoWrapper.kt
@@ -37,6 +37,7 @@ class SpeedLimitedRepoWrapper(
         filename: String,
         info: ArchiveFileInfo,
         stream: InputStream,
+        monitor: (copiedBytes: Long) -> Unit,
     ) = delayed(200) {
         base.save(filename, info, stream)
     }